change import signature, move `ImportPath` to graveler and refactor accodingly
