+++ pkg/graveler/committed/manager.go
@@ -294,6 +294,6 @@ func (c *committedManager) GetRangeIDByKey(ctx context.Context, ns graveler.Stor
 	return graveler.RangeID(r.ID), nil
 }
 
-func (c *committedManager) Import(ctx context.Context, ns graveler.StorageNamespace, destination, source graveler.MetaRangeID) (graveler.MetaRangeID, error) {
+func (c *committedManager) Import(ctx context.Context, ns graveler.StorageNamespace, destination, source graveler.MetaRangeID, importPaths []graveler.ImportPath) (graveler.MetaRangeID, error) {
 	return "", nil
 }
