+++ pkg/graveler/testutil/fakes.go
@@ -118,7 +118,7 @@ func (c *CommittedFake) GetRange(_ context.Context, _ graveler.StorageNamespace,
 	return graveler.RangeAddress(fmt.Sprintf("fake://prefix/%s(range)", rangeID)), nil
 }
 
-func (c *CommittedFake) Import(ctx context.Context, ns graveler.StorageNamespace, destination, source graveler.MetaRangeID) (graveler.MetaRangeID, error) {
+func (c *CommittedFake) Import(ctx context.Context, ns graveler.StorageNamespace, destination, source graveler.MetaRangeID, importPaths []graveler.ImportPath) (graveler.MetaRangeID, error) {
 	return "", nil
 }
